fea: Register command for the plugin. Rebuild the structure. The command feature have been tested yet.
